// fe/src/lib/api/generated/model/designationCode.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type DesignationCode = typeof DesignationCode[keyof typeof DesignationCode];


export const DesignationCode = {
  DPO: 'DPO',
  HEAD_GUIDANCE_COUNSELOR: 'HEAD_GUIDANCE_COUNSELOR',
} as const;