// fe/src/lib/api/generated/model/accountDetailResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { DesignationCode } from './designationCode';
import type { RoleCode } from './roleCode';
import type { StudentLifecycleCode } from './studentLifecycleCode';

export interface AccountDetailResponse {
  created_at: string;
  designations: DesignationCode[];
  email: string;
  email_verified: boolean;
  email_verified_at: string | null;
  first_name: string;
  full_name: string;
  id: string;
  institutional_id: string | null;
  is_active: boolean;
  last_name: string;
  mfa_enabled: boolean;
  middle_name: string;
  password_configured: boolean;
  role: RoleCode;
  student_lifecycle_status: StudentLifecycleCode | null;
  suffix: string;
  updated_at: string;
}